Finished windchill

+++ f_to_c.py
@@ -3,7 +3,7 @@ Takes user inputted number in Fahrenheit and outputs it as Celsius using the giv
 """
 
 #input
-degree = int(input("What is the temperature in Fahrenheit: "))
+degree = float(input("What is the temperature in Fahrenheit: "))
 
 #output with formula
 print("The temperature in Celsius is ", (degree - 32)*5/9)
